Update Logstash tutorial to reference integration plugin
The Logstash input and output plugins for Kafka are now provided by
a single "integration" plugin, and the previous stand-alone plugins are
no longer maintained. By including instructions for the integration we
encourage our users to get the latest version that includes fixes and
updates.

Integration plugins require Logstash 6.5+ (released 2018-11), and
the oldest non-EOL version of Logstash as of this writing is 7.17.